Ajuste na organizacao do projeto, adicionado interfaces
