+++ src/main/java/br/com/qrcodegeneration/QrcodeGenerationApplication.java
@@ -1,12 +1,18 @@
 package br.com.qrcodegeneration;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+
 
 @SpringBootApplication
 public class QrcodeGenerationApplication {
 
+	private static Logger logger = LoggerFactory.getLogger(QrcodeGenerationApplication.class);
+
 	public static void main(String[] args) {
+		logger.info("Iniciando servico de geracao de QR-Code");
 		SpringApplication.run(QrcodeGenerationApplication.class, args);
 	}
